Adding StageTrait
If you think about it…
Data comes from a lot of places. I guess that’s why PHP made a
`$_REQUEST` variable.  This one mashes up the `$_COOKIE`, `$_SESSION`,
`$_GET`, `$_POST` and eventually the URL parameters together, so you
don’t need to look in several places for data. I also considered that
`$_SESSION`s should be soft insert meaning that if you are updating
your profile, it won’t overwrite with the current Session data. and For
security we force that route parameters overwrite all.